requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
httpx==0.19.0
aiofiles==0.7.0
pytesseract==0.3.8
Pillow==8.3.2
spacy==3.1.3
PyPDF2==1.26.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==0.19.0
aiohttp==3.8.1
pydantic==1.8.2